Name servlets following convention.  Add to web.xml so servlets are registered to an app URL.
