+++ src/crud/TinyOverwrite.java
@@ -19,14 +19,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import net.sf.json.JSONObject;
 import tokens.TinyTokenManager;
-import java.util.List;
-import java.util.Map;
+
 
 /**
  *
  * @author bhaberbe
  */
-public class tinySave extends HttpServlet {
+public class TinyOverwrite extends HttpServlet {
+    //private final TinyTokenManager manager = new TinyTokenManager("E:\\tinyThings\\Source Packages\\tiny.properties");
+
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -36,21 +37,20 @@ public class tinySave extends HttpServlet {
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
-       protected void processRequest(HttpServletRequest request, HttpServletResponse response)
+        protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException, Exception {
-        System.out.println("Tiny save...");
-        request.setCharacterEncoding("UTF-8");
         response.setHeader("Content-Type", "application/json; charset=utf-8");
         response.setCharacterEncoding("UTF-8");
-        String line;
-        StringBuilder sb = new StringBuilder();
-        int codeOverwrite = 500;
         TinyTokenManager manager = new TinyTokenManager();
         BufferedReader bodyReader = request.getReader();
         StringBuilder bodyString = new StringBuilder();
+        String line;
+        StringBuilder sb = new StringBuilder();
+        int codeOverwrite = 500;
         JSONObject requestJSON = new JSONObject();
         String requestString;
         boolean moveOn = false;
+        //Gather user provided parameters from BODY of request, not parameters
         while ((line = bodyReader.readLine()) != null)
         {
           bodyString.append(line);
@@ -64,22 +64,22 @@ public class tinySave extends HttpServlet {
         catch(Exception ex){
             response.setStatus(500);
             response.getWriter().print(ex);
-        }       
+        }
         //If it was JSON
         if(moveOn){
+            //Get public token for requests from property file
             String pubTok = manager.getAccessToken();
             boolean expired = manager.checkTokenExpiry();
             if(expired){
                 System.out.println("Tiny thing detected an expired token, auto getting and setting a new one...");
                 pubTok = manager.generateNewAccessToken();
             }
-            System.out.println("Bearer token is set for Tiny Save, connecting to RERUM for create...");
             //Point to rerum server v1
-            URL postUrl = new URL(Constant.RERUM_API_ADDR + "/create.action");
+            URL postUrl = new URL(Constant.RERUM_API_ADDR + "/overwrite.action");
             HttpURLConnection connection = (HttpURLConnection) postUrl.openConnection();
             connection.setDoOutput(true);
             connection.setDoInput(true);
-            connection.setRequestMethod("POST");
+            connection.setRequestMethod("PUT");
             connection.setUseCaches(false);
             connection.setInstanceFollowRedirects(true);
             connection.setRequestProperty("Content-Type", "application/json; charset=utf-8");
@@ -87,10 +87,8 @@ public class tinySave extends HttpServlet {
             connection.connect();
             try{
                 DataOutputStream out = new DataOutputStream(connection.getOutputStream());
-                byte[] toWrite = requestJSON.toString().getBytes("UTF-8");
                 //Pass in the user provided JSON for the body of the rerumserver v1 request
-                //out.writeBytes(requestJSON.toString()); //something about this is breaking special chars when using this as am open endpoint API
-                out.write(toWrite);
+                out.writeBytes(requestJSON.toString());
                 out.flush();
                 out.close(); 
                 codeOverwrite = connection.getResponseCode();
@@ -100,13 +98,6 @@ public class tinySave extends HttpServlet {
                     sb.append(line);
                 }
                 reader.close();
-                for (Map.Entry<String, List<String>> entries : connection.getHeaderFields().entrySet()) {
-                    String values = "";
-                    for (String value : entries.getValue()) {
-                        values += value + ",";
-                    }
-                    response.setHeader(entries.getKey(), values);
-                }
             }
             catch(IOException ex){
                 //Need to get the response RERUM sent back.
@@ -118,20 +109,14 @@ public class tinySave extends HttpServlet {
                 error.close();
             }
             connection.disconnect();
-            System.out.println("RERUM create responded, out that to user!");
-            //Hand back rerumserver response as this API's response.
-            if(manager.getAPISetting().equals("true")){
-                response.addHeader("Access-Control-Allow-Origin", "*"); //To use this as an API, it must contain CORS headers
-            }
             response.setStatus(codeOverwrite);
             response.getWriter().print(sb.toString());
         }
         
     }
 
-    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
     /**
-     * Handles the HTTP <code>GET</code> method.
+     * Handles the HTTP <code>PUT</code> method.
      *
      * @param request servlet request
      * @param response servlet response
@@ -139,59 +124,12 @@ public class tinySave extends HttpServlet {
      * @throws IOException if an I/O error occurs
      */
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
+    protected void doPut(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         try {
             processRequest(request, response);
         } catch (Exception ex) {
-            Logger.getLogger(tinySave.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
-    /**
-     * Handles the HTTP <code>POST</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        try {
-            processRequest(request, response);
-        } catch (Exception ex) {
-            Logger.getLogger(tinySave.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-    
-    /**
-     * Handles the HTTP <code>OPTIONS</code> preflight method.
-     * This should be a configurable option.  Turning this on means you
-     * intend for this version of Tiny Things to work like an open API.  
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
-    @Override
-    protected void doOptions(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        try {
-            TinyTokenManager manager = new TinyTokenManager();
-            String openAPI = manager.getAPISetting();
-            if(openAPI.equals("true")){
-                //These headers must be present to pass browser preflight for CORS
-                response.addHeader("Access-Control-Allow-Origin", "*");
-                response.addHeader("Access-Control-Allow-Headers", "*");
-                response.addHeader("Access-Control-Allow-Methods", "*");
-            }
-            response.setStatus(200);
-            
-        } catch (Exception ex) {
-            Logger.getLogger(tinyQuery.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(TinyOverwrite.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 
@@ -202,7 +140,7 @@ public class tinySave extends HttpServlet {
      */
     @Override
     public String getServletInfo() {
-        return "Short description";
+        return "Overwrites an object in the RERUM store without versioning.";
     }// </editor-fold>
 
 }
