+++ src/crud/TinySave.java
@@ -22,14 +22,11 @@ import tokens.TinyTokenManager;
 import java.util.List;
 import java.util.Map;
 
-
 /**
  *
  * @author bhaberbe
  */
-public class tinyUpdate extends HttpServlet {
-    //private final TinyTokenManager manager = new TinyTokenManager("E:\\tinyThings\\Source Packages\\tiny.properties");
-
+public class TinySave extends HttpServlet {
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -39,21 +36,21 @@ public class tinyUpdate extends HttpServlet {
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
-        protected void processRequest(HttpServletRequest request, HttpServletResponse response)
+       protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException, Exception {
+        System.out.println("Tiny save...");
         request.setCharacterEncoding("UTF-8");
         response.setHeader("Content-Type", "application/json; charset=utf-8");
         response.setCharacterEncoding("UTF-8");
-        TinyTokenManager manager = new TinyTokenManager();
-        BufferedReader bodyReader = request.getReader();
-        StringBuilder bodyString = new StringBuilder();
         String line;
         StringBuilder sb = new StringBuilder();
         int codeOverwrite = 500;
+        TinyTokenManager manager = new TinyTokenManager();
+        BufferedReader bodyReader = request.getReader();
+        StringBuilder bodyString = new StringBuilder();
         JSONObject requestJSON = new JSONObject();
         String requestString;
         boolean moveOn = false;
-        //Gather user provided parameters from BODY of request, not parameters
         while ((line = bodyReader.readLine()) != null)
         {
           bodyString.append(line);
@@ -67,22 +64,22 @@ public class tinyUpdate extends HttpServlet {
         catch(Exception ex){
             response.setStatus(500);
             response.getWriter().print(ex);
-        }
+        }       
         //If it was JSON
         if(moveOn){
-            //Get public token for requests from property file
             String pubTok = manager.getAccessToken();
             boolean expired = manager.checkTokenExpiry();
             if(expired){
                 System.out.println("Tiny thing detected an expired token, auto getting and setting a new one...");
                 pubTok = manager.generateNewAccessToken();
             }
+            System.out.println("Bearer token is set for Tiny Save, connecting to RERUM for create...");
             //Point to rerum server v1
-            URL postUrl = new URL(Constant.RERUM_API_ADDR + "/update.action");
+            URL postUrl = new URL(Constant.RERUM_API_ADDR + "/create.action");
             HttpURLConnection connection = (HttpURLConnection) postUrl.openConnection();
             connection.setDoOutput(true);
             connection.setDoInput(true);
-            connection.setRequestMethod("PUT");
+            connection.setRequestMethod("POST");
             connection.setUseCaches(false);
             connection.setInstanceFollowRedirects(true);
             connection.setRequestProperty("Content-Type", "application/json; charset=utf-8");
@@ -92,7 +89,7 @@ public class tinyUpdate extends HttpServlet {
                 DataOutputStream out = new DataOutputStream(connection.getOutputStream());
                 byte[] toWrite = requestJSON.toString().getBytes("UTF-8");
                 //Pass in the user provided JSON for the body of the rerumserver v1 request
-                //out.writeBytes(requestJSON.toString());
+                //out.writeBytes(requestJSON.toString()); //something about this is breaking special chars when using this as am open endpoint API
                 out.write(toWrite);
                 out.flush();
                 out.close(); 
@@ -110,7 +107,8 @@ public class tinyUpdate extends HttpServlet {
                     }
                     response.setHeader(entries.getKey(), values);
                 }
-            } catch (IOException ex) {
+            }
+            catch(IOException ex){
                 //Need to get the response RERUM sent back.
                 BufferedReader error = new BufferedReader(new InputStreamReader(connection.getErrorStream(),"utf-8"));
                 String errorLine = "";
@@ -120,6 +118,8 @@ public class tinyUpdate extends HttpServlet {
                 error.close();
             }
             connection.disconnect();
+            System.out.println("RERUM create responded, out that to user!");
+            //Hand back rerumserver response as this API's response.
             if(manager.getAPISetting().equals("true")){
                 response.addHeader("Access-Control-Allow-Origin", "*"); //To use this as an API, it must contain CORS headers
             }
@@ -144,12 +144,12 @@ public class tinyUpdate extends HttpServlet {
         try {
             processRequest(request, response);
         } catch (Exception ex) {
-            Logger.getLogger(tinyUpdate.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(TinySave.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 
     /**
-     * Handles the HTTP <code>PUT</code> method.
+     * Handles the HTTP <code>POST</code> method.
      *
      * @param request servlet request
      * @param response servlet response
@@ -157,16 +157,16 @@ public class tinyUpdate extends HttpServlet {
      * @throws IOException if an I/O error occurs
      */
     @Override
-    protected void doPut(HttpServletRequest request, HttpServletResponse response)
+    protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         try {
             processRequest(request, response);
         } catch (Exception ex) {
-            Logger.getLogger(tinyUpdate.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(TinySave.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
     
-   /**
+    /**
      * Handles the HTTP <code>OPTIONS</code> preflight method.
      * This should be a configurable option.  Turning this on means you
      * intend for this version of Tiny Things to work like an open API.  
@@ -191,7 +191,7 @@ public class tinyUpdate extends HttpServlet {
             response.setStatus(200);
             
         } catch (Exception ex) {
-            Logger.getLogger(tinyQuery.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(TinyQuery.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
